+++ kernel/include/kernel/pmm.h
@@ -24,8 +24,8 @@ typedef struct page_t {
     uint8_t flags; // indicate status of the page
     uint8_t level; // indicate size: 2^level after it belongs to it
     uint32_t index; // indicate relative index in pages array
+    uint32_t prev_page; //indicate index of previous unallocate pages block in the free list
     uint32_t next_page; //indicate index of next unallocate pages block in the free list
-    uint32_t addr; // physic address
 } __attribute__((packed)) page_t;
 
 
@@ -34,20 +34,15 @@ typedef struct normal_mem_t {
     uint32_t length; // size in bytes
     uint32_t pages_number; // size in pages
     uint32_t free_lists[BUDDY_MAXLEVEL];  // free list
-    page_t  pages[MAX_PAGES_NUM]; // pages table
+    // page_t  pages[MAX_PAGES_NUM]; // pages table, fixed
+    page_t*  pages; // pages table, dynamic
 } __attribute__((packed)) normal_mem_t;
 
-
-
-
-
-
-
+void print_free_buddy_blocks();
 
 // start and end pisition in memory of kernel file
 extern uint8_t kernel_start[];
 extern uint8_t kernel_end[];
-
 
 
 // print physical memory map
@@ -55,6 +50,30 @@ void show_memory_map();
 
 // init normal memory zone
 void init_normal_mem_zone();
+
+// allocate physic memory
+void* kmalloc(uint32_t byte_size);
+
+// allocate pages
+uint32_t palloc(uint32_t page_size);
+
+
+// free physic memory
+uint32_t kfree(void* addr);
+
+
+inline uint32_t is_power_of_2 (uint32_t n);
+inline uint32_t next_power_of_2 (uint32_t n);
+inline uint32_t power_of_2 (uint32_t n);
+
+// buddy_system
+
+/** 
+ * return first page index
+ * */
+uint32_t buddy_system_get_page_block(uint32_t level);
+
+uint32_t buddy_system_free_page_block(uint32_t page_index);
 
 
 #endif // INCLUDE_PMM_H