+++ qemu.sh
@@ -2,4 +2,4 @@
 set -e
 . ./iso.sh
 
-qemu-system-$(./target-triplet-to-arch.sh $HOST) -cdrom chaosos.iso
+qemu-system-$(./target-triplet-to-arch.sh $HOST) -m 128 -cdrom chaosos.iso