+++ CHANGELOG.md
@@ -9,7 +9,17 @@
     - impelements in kernel/intterrupt.c
 - isr for ISR0 to ISR15
 - a timer interrupt handler entry which can be assigned by user
+- tty cursor move
+- add surport for ```%x``` and ```%X```
+- pages table
+- physic memory management: buddy system
+    - kmalloc(size_t size_in_bytes) : acquire a contigious physic memory
+    - kfree(void* pointer) : free a allocated memory from kmalloc 
+- support for any memory size less than 4 GiB
 
+### Fixed
+- bug: int tty.c, output bug of terminal_roll()
+}
 ## [0.0.1] - 2020-06-22
 ### Added
 - kernel skeleton